Refactor: Organize sidebar components into dedicated directory
- Move primary-sidebar.tsx and project-sidebar.tsx to components/sidebars/
- Update import paths in app/settings/layout.tsx and app/projects/[id]/layout.tsx
- Improve code organization and modularity
